Define `extract_package_names_from_specs` in `api/utils.cpp`
