+++ libmamba/src/core/util.cpp
@@ -1725,22 +1725,5 @@ namespace mamba
         return copy;
     }
 
-    std::vector<std::string> extract_package_names_from_specs(const std::vector<std::string>& specs)
-    {
-        std::vector<std::string> names;
-        for (const auto& s : specs)
-        {
-            auto parsed = specs::MatchSpec::parse(s);
-            if (parsed && !parsed->name().is_free())
-            {
-                auto name = parsed->name().to_string();
-                if (!name.empty())
-                {
-                    names.push_back(std::move(name));
-                }
-            }
-        }
-        return names;
-    }
 
 }  // namespace mamba
